+++ src/imio/smartweb/core/contents/pages/page/content.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 
+from collective.instancebehavior.interfaces import IInstanceBehaviorAssignableContent
 from imio.smartweb.core.contents import IPages
 from imio.smartweb.core.contents import Pages
 from zope.interface import implementer
@@ -9,7 +10,7 @@ class IPage(IPages):
     """Marker interface and Dexterity Python Schema for Page"""
 
 
-@implementer(IPage)
+@implementer(IPage, IInstanceBehaviorAssignableContent)
 class Page(Pages):
     """Page class"""